requirements.txt
streamlit==1.24.1
openai==0.27.0
pandas==1.5.3
python-dotenv==1.0.0
pydantic==1.10.7
